refactor: Remove unused code and rename variables for in-app wallet
