+++ packages/thirdweb/src/react/web/ui/ConnectWallet/Details.tsx
@@ -49,21 +49,9 @@ import type {
   ConnectButton_detailsButtonOptions,
   ConnectButton_detailsModalOptions,
 } from "./ConnectWalletProps.js";
-// import { connectionManager } from "../../../core/connectionManager.js";
 import { SendFunds } from "./screens/SendFunds.js";
 import { ReceiveFunds } from "./screens/ReceiveFunds.js";
 import type { Chain } from "../../../../chains/types.js";
-// import type {
-//   Wallet,
-//   WalletWithPersonalAccount,
-// } from "../../../../wallets/interfaces/wallet.js";
-// import {
-//   // personalAccountToSmartAccountMap,
-//   smartWalletMetadata,
-// } from "../../../../wallets/smart/index._ts";
-// import type { EmbeddedWallet } from "../../../../wallets/in-app/core/wallet/index._ts";
-// import { localWalletMetadata } from "../../../../wallets/local/index._ts";
-// import { ExportLocalWallet } from "./screens/ExportLocalWallet._tsx";
 import { swapTransactionsStore } from "./screens/Buy/swap/pendingSwapTx.js";
 import { SwapScreen } from "./screens/Buy/SwapScreen.js";
 import { SwapTransactionsScreen } from "./screens/SwapTransactionsScreen.js";
@@ -324,7 +312,7 @@ export const ConnectedWalletDetails: React.FC<{
 
       <Container px="lg">
         <ConnectedToSmartWallet />
-        <EmbeddedWalletEmail />
+        <InAppWalletEmail />
 
         {/* Send, Receive, Swap */}
         <Container
@@ -815,10 +803,10 @@ function ConnectedToSmartWallet() {
   return null;
 }
 
-function EmbeddedWalletEmail() {
+function InAppWalletEmail() {
   const { client } = useWalletConnectionCtx();
   const emailQuery = useQuery({
-    queryKey: ["embedded-wallet-user", client],
+    queryKey: ["in-app-wallet-user", client],
     queryFn: async () => {
       const data = await getUserEmail({
         client: client,
